+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * redis: support HMGET
 * db: updated upsert SQL with latest syntax for MySQL
   > refer to https://dev.mysql.com/doc/refman/8.4/en/insert-on-duplicate.html
+* db: tweak postgres query analyzer rules
 
 ### 9.4.2 (1/26/26 - 2/15/26)